WAT Setup File executable
Hi, this is Word Association test executable file for your pc or laptop, it should be x64 architecture. While running the program, click on start and every 15 seconds, the word on the screen changes and you have to write a sentence on the word in 15 seconds. It works, I have tried it. While exiting the program, click on quit button, it will take sometime to quit as this is a basic application and does not use threading.